requirements.txt
GroupyAPI==0.5.8
mock==1.0.1
Pillow==2.5.3
requests==2.3.0
responses==0.3.0
six==1.9.0
wxPython-Phoenix==3.0.3.dev1712+dcd5d1d
